+++ src/main/java/uk/q3c/krail/i18n/bind/I18NModule.java
@@ -33,7 +33,7 @@ import uk.q3c.krail.i18n.persist.source.DefaultPatternSource;
 import uk.q3c.krail.i18n.persist.source.DefaultPatternSourceProvider;
 import uk.q3c.krail.i18n.translate.DefaultTranslate;
 import uk.q3c.krail.i18n.util.DefaultPatternUtility;
-import uk.q3c.krail.persist.KrailPersistenceUnitHelper;
+import uk.q3c.krail.persist.I18NPersistenceHelper;
 import uk.q3c.util.clazz.DefaultUnenhancedClassIdentifier;
 import uk.q3c.util.clazz.UnenhancedClassIdentifier;
 
@@ -55,7 +55,7 @@ import static com.google.common.base.Preconditions.*;
 
 public class I18NModule extends AbstractModule {
 
-    private final TypeLiteral<Class<? extends Annotation>> annotationLiteral = KrailPersistenceUnitHelper.annotationClassLiteral();
+    private final TypeLiteral<Class<? extends Annotation>> annotationLiteral = I18NPersistenceHelper.annotationClassLiteral();
     private final TypeLiteral<PatternDao> patternDaoTypeLiteral = new TypeLiteral<PatternDao>() {
     };
     private Locale defaultLocale = Locale.UK;